Split TensorDict helper functions
