+++ tensordict/_td.py
@@ -5088,473 +5088,18 @@ def _update_metadata(*, metadata, key, value, is_collection):
         }
 
 
-def from_module(
-    module,
-    as_module: bool = False,
-    lock: bool = True,
-    use_state_dict: bool = False,
-):
-    """Copies the params and buffers of a module in a tensordict.
-
-    Args:
-        module (nn.Module): the module to get the parameters from.
-        as_module (bool, optional): if ``True``, a :class:`~tensordict.nn.TensorDictParams`
-            instance will be returned which can be used to store parameters
-            within a :class:`torch.nn.Module`. Defaults to ``False``.
-        lock (bool, optional): if ``True``, the resulting tensordict will be locked.
-            Defaults to ``True``.
-        use_state_dict (bool, optional): if ``True``, the state-dict from the
-            module will be used and unflattened into a TensorDict with
-            the tree structure of the model. Defaults to ``False``.
-
-            .. note::
-                This is particularly useful when state-dict hooks have to be used.
-
-    Examples:
-        >>> from torch import nn
-        >>> module = nn.TransformerDecoder(
-        ...     decoder_layer=nn.TransformerDecoderLayer(nhead=4, d_model=4),
-        ...     num_layers=1)
-        >>> params = from_module(module)
-        >>> print(params["layers", "0", "linear1"])
-        TensorDict(
-            fields={
-                bias: Parameter(shape=torch.Size([2048]), device=cpu, dtype=torch.float32, is_shared=False),
-                weight: Parameter(shape=torch.Size([2048, 4]), device=cpu, dtype=torch.float32, is_shared=False)},
-            batch_size=torch.Size([]),
-            device=None,
-            is_shared=False)
-    """
-    return TensorDict.from_module(
-        module=module, as_module=as_module, lock=lock, use_state_dict=use_state_dict
-    )
-
-
-def from_modules(
-    *modules,
-    as_module: bool = False,
-    lock: bool = True,
-    use_state_dict: bool = False,
-    lazy_stack: bool = False,
-    expand_identical: bool = False,
-):
-    """Retrieves the parameters of several modules for ensebmle learning/feature of expects applications through vmap.
-
-    Args:
-        modules (sequence of nn.Module): the modules to get the parameters from.
-            If the modules differ in their structure, a lazy stack is needed
-            (see the ``lazy_stack`` argument below).
-
-    Keyword Args:
-        as_module (bool, optional): if ``True``, a :class:`~tensordict.nn.TensorDictParams`
-            instance will be returned which can be used to store parameters
-            within a :class:`torch.nn.Module`. Defaults to ``False``.
-        lock (bool, optional): if ``True``, the resulting tensordict will be locked.
-            Defaults to ``True``.
-        use_state_dict (bool, optional): if ``True``, the state-dict from the
-            module will be used and unflattened into a TensorDict with
-            the tree structure of the model. Defaults to ``False``.
-
-            .. note::
-                This is particularly useful when state-dict hooks have to be used.
-
-        lazy_stack (bool, optional): whether parameters should be densly or
-            lazily stacked. Defaults to ``False`` (dense stack).
-
-            .. note::
-                ``lazy_stack`` and ``as_module`` are exclusive features.
-
-            .. warning::
-                There is a crucial difference between lazy and non-lazy outputs
-                in that non-lazy output will reinstantiate parameters with the
-                desired batch-size, while ``lazy_stack`` will just represent
-                the parameters as lazily stacked. This means that whilst the
-                original parameters can safely be passed to an optimizer
-                when ``lazy_stack=True``, the new parameters need to be passed
-                when it is set to ``True``.
-
-            .. warning::
-                Whilst it can be tempting to use a lazy stack to keep the
-                orignal parameter references, remember that lazy stack
-                perform a stack each time :meth:`~.get` is called. This will
-                require memory (N times the size of the parameters, more if a
-                graph is built) and time to be computed.
-                It also means that the optimizer(s) will contain more
-                parameters, and operations like :meth:`~torch.optim.Optimizer.step`
-                or :meth:`~torch.optim.Optimizer.zero_grad` will take longer
-                to be executed. In general, ``lazy_stack`` should be reserved
-                to very few use cases.
-
-        expand_identical (bool, optional): if ``True`` and the same parameter (same
-            identity) is being stacked to itself, an expanded version of this parameter
-            will be returned instead. This argument is ignored when ``lazy_stack=True``.
-
-    Examples:
-        >>> from torch import nn
-        >>> from tensordict import from_modules
-        >>> torch.manual_seed(0)
-        >>> empty_module = nn.Linear(3, 4, device="meta")
-        >>> n_models = 2
-        >>> modules = [nn.Linear(3, 4) for _ in range(n_models)]
-        >>> params = from_modules(*modules)
-        >>> print(params)
-        TensorDict(
-            fields={
-                bias: Parameter(shape=torch.Size([2, 4]), device=cpu, dtype=torch.float32, is_shared=False),
-                weight: Parameter(shape=torch.Size([2, 4, 3]), device=cpu, dtype=torch.float32, is_shared=False)},
-            batch_size=torch.Size([2]),
-            device=None,
-            is_shared=False)
-        >>> # example of batch execution
-        >>> def exec_module(params, x):
-        ...     with params.to_module(empty_module):
-        ...         return empty_module(x)
-        >>> x = torch.randn(3)
-        >>> y = torch.vmap(exec_module, (0, None))(params, x)
-        >>> assert y.shape == (n_models, 4)
-        >>> # since lazy_stack = False, backprop leaves the original params untouched
-        >>> y.sum().backward()
-        >>> assert params["weight"].grad.norm() > 0
-        >>> assert modules[0].weight.grad is None
-
-    With ``lazy_stack=True``, things are slightly different:
-
-        >>> params = TensorDict.from_modules(*modules, lazy_stack=True)
-        >>> print(params)
-        LazyStackedTensorDict(
-            fields={
-                bias: Tensor(shape=torch.Size([2, 4]), device=cpu, dtype=torch.float32, is_shared=False),
-                weight: Tensor(shape=torch.Size([2, 4, 3]), device=cpu, dtype=torch.float32, is_shared=False)},
-            exclusive_fields={
-            },
-            batch_size=torch.Size([2]),
-            device=None,
-            is_shared=False,
-            stack_dim=0)
-        >>> # example of batch execution
-        >>> y = torch.vmap(exec_module, (0, None))(params, x)
-        >>> assert y.shape == (n_models, 4)
-        >>> y.sum().backward()
-        >>> assert modules[0].weight.grad is not None
-
-
-    """
-    return TensorDict.from_modules(
-        *modules,
-        lazy_stack=lazy_stack,
-        expand_identical=expand_identical,
-        lock=lock,
-        use_state_dict=use_state_dict,
-        as_module=as_module,
-    )
-
-
-def from_pytree(
-    pytree,
-    *,
-    batch_size: torch.Size | None = None,
-    auto_batch_size: bool = False,
-    batch_dims: int | None = None,
-):
-    """Converts a pytree to a TensorDict instance.
-
-    This method is designed to keep the pytree nested structure as much as possible.
-
-    Additional non-tensor keys are added to keep track of each level's identity, providing
-    a built-in pytree-to-tensordict bijective transform API.
-
-    Accepted classes currently include lists, tuples, named tuples and dict.
-
-    .. note::
-        For dictionaries, non-NestedKey keys are registered separately as :class:`~tensordict.NonTensorData`
-        instances.
-
-    .. note::
-        Tensor-castable types (such as int, float or np.ndarray) will be converted to torch.Tensor instances.
-        Note that this transformation is surjective: transforming back the tensordict to a pytree will not
-        recover the original types.
-
-    Examples:
-        >>> # Create a pytree with tensor leaves, and one "weird"-looking dict key
-        >>> class WeirdLookingClass:
-        ...     pass
-        ...
-        >>> weird_key = WeirdLookingClass()
-        >>> # Make a pytree with tuple, lists, dict and namedtuple
-        >>> pytree = (
-        ...     [torch.randint(10, (3,)), torch.zeros(2)],
-        ...     {
-        ...         "tensor": torch.randn(
-        ...             2,
-        ...         ),
-        ...         "td": TensorDict({"one": 1}),
-        ...         weird_key: torch.randint(10, (2,)),
-        ...         "list": [1, 2, 3],
-        ...     },
-        ...     {"named_tuple": TensorDict({"two": torch.ones(1) * 2}).to_namedtuple()},
-        ... )
-        >>> # Build a TensorDict from that pytree
-        >>> td = from_pytree(pytree)
-        >>> # Recover the pytree
-        >>> pytree_recon = td.to_pytree()
-        >>> # Check that the leaves match
-        >>> def check(v1, v2):
-        >>>     assert (v1 == v2).all()
-        >>>
-        >>> torch.utils._pytree.tree_map(check, pytree, pytree_recon)
-        >>> assert weird_key in pytree_recon[1]
-
-    """
-    return TensorDict.from_pytree(
-        pytree,
-        batch_size=batch_size,
-        auto_batch_size=auto_batch_size,
-        batch_dims=batch_dims,
-    )
-
-
-def stack(input: Sequence[T], dim: int = 0, *, out=None) -> T:
-    """Stacks tensordicts into a single tensordict along the given dimension.
-
-    This call is equivalent to calling :func:`torch.stack` but is compatible with torch.compile.
-
-    """
-    return torch.stack(input, dim=dim, out=out)
-
-
-def lazy_stack(input: Sequence[T], dim: int = 0, *, out=None) -> T:
-    """Creates a lazy stack of tensordicts.
-
-    See :meth:`~tensordict.LazyStackTensorDict.lazy_stack` for details.
-    """
-    return TensorDict.lazy_stack(input, dim=dim, out=out)
-
-
-def cat(input: Sequence[T], dim: int = 0, *, out=None) -> T:
-    """Concatenates tensordicts into a single tensordict along the given dimension.
-
-    This call is equivalent to calling :func:`torch.cat` but is compatible with torch.compile.
-
-    """
-    return torch.cat(input, dim=dim, out=out)
-
-
-def maybe_dense_stack(input: Sequence[T], dim: int = 0, *, out=None, **kwargs) -> T:
-    """Attempts to make a dense stack of tensordicts, and falls back on lazy stack when required..
-
-    See :meth:`~tensordict.LazyStackTensorDict.maybe_dense_stack` for details.
-    """
-    return TensorDict.maybe_dense_stack(input, dim=dim, out=out, **kwargs)
-
-
-def fromkeys(keys: List[NestedKey], value: Any = 0):
-    """Creates a tensordict from a list of keys and a single value.
-
-    Args:
-        keys (list of NestedKey): An iterable specifying the keys of the new dictionary.
-        value (compatible type, optional): The value for all keys. Defaults to ``0``.
-    """
-    return TensorDict.fromkeys(keys=keys, value=value)
-
-
-def from_consolidated(filename):
-    """Reconstructs a tensordict from a consolidated file."""
-    return TensorDict.from_consolidated(filename)
-
-
-def load(
-    prefix: str | Path,
-    device: torch.device | None = None,
-    non_blocking: bool = False,
-    *,
-    out: TensorCollection | None = None,
-    robust_key: bool | None = None,
-) -> Self:
-    """Loads a tensordict from disk.
-
-    This class method is a proxy to :meth:`~.load_memmap`.
-    """
-    return load_memmap(
-        prefix=prefix,
-        device=device,
-        non_blocking=non_blocking,
-        out=out,
-        robust_key=robust_key,
-    )
-
-
-def load_memmap(
-    prefix: str | Path,
-    device: torch.device | None = None,
-    non_blocking: bool = False,
-    *,
-    out: TensorCollection | None = None,
-    robust_key: bool | None = None,
-) -> Self:
-    """Loads a memory-mapped tensordict from disk.
-
-    Args:
-        prefix (str or Path to folder): the path to the folder where the
-            saved tensordict should be fetched.
-        device (torch.device or equivalent, optional): if provided, the
-            data will be asynchronously cast to that device.
-            Supports `"meta"` device, in which case the data isn't loaded
-            but a set of empty "meta" tensors are created. This is
-            useful to get a sense of the total model size and structure
-            without actually opening any file.
-        non_blocking (bool, optional): if ``True``, synchronize won't be
-            called after loading tensors on device. Defaults to ``False``.
-        out (TensorDictBase, optional): optional tensordict where the data
-            should be written.
-        robust_key (bool, optional): if ``True``, expects robust key encoding was used
-            when saving and decodes filenames accordingly. If ``False``, uses legacy
-            behavior. If ``None`` (default), emits a deprecation warning and falls
-            back to legacy behavior. Will default to ``True`` in v0.12.
-
-    Examples:
-        >>> from tensordict import TensorDict, load_memmap
-        >>> td = TensorDict.fromkeys(["a", "b", "c", ("nested", "e")], 0)
-        >>> td.memmap("./saved_td")
-        >>> td_load = TensorDict.load_memmap("./saved_td")
-        >>> assert (td == td_load).all()
-
-    This method also allows loading nested tensordicts.
-
-    Examples:
-        >>> nested = TensorDict.load_memmap("./saved_td/nested")
-        >>> assert nested["e"] == 0
-
-    A tensordict can also be loaded on "meta" device or, alternatively,
-    as a fake tensor.
-
-    Examples:
-        >>> import tempfile
-        >>> td = TensorDict({"a": torch.zeros(()), "b": {"c": torch.zeros(())}})
-        >>> with tempfile.TemporaryDirectory() as path:
-        ...     td.save(path)
-        ...     td_load = load_memmap(path, device="meta")
-        ...     print("meta:", td_load)
-        ...     from torch._subclasses import FakeTensorMode
-        ...     with FakeTensorMode():
-        ...         td_load = load_memmap(path)
-        ...         print("fake:", td_load)
-        meta: TensorDict(
-            fields={
-                a: Tensor(shape=torch.Size([]), device=meta, dtype=torch.float32, is_shared=False),
-                b: TensorDict(
-                    fields={
-                        c: Tensor(shape=torch.Size([]), device=meta, dtype=torch.float32, is_shared=False)},
-                    batch_size=torch.Size([]),
-                    device=meta,
-                    is_shared=False)},
-            batch_size=torch.Size([]),
-            device=meta,
-            is_shared=False)
-        fake: TensorDict(
-            fields={
-                a: FakeTensor(shape=torch.Size([]), device=cpu, dtype=torch.float32, is_shared=False),
-                b: TensorDict(
-                    fields={
-                        c: FakeTensor(shape=torch.Size([]), device=cpu, dtype=torch.float32, is_shared=False)},
-                    batch_size=torch.Size([]),
-                    device=cpu,
-                    is_shared=False)},
-            batch_size=torch.Size([]),
-            device=cpu,
-            is_shared=False)
-
-    """
-    return TensorDict.load_memmap(
-        prefix=prefix,
-        device=device,
-        non_blocking=non_blocking,
-        out=out,
-        robust_key=robust_key,
-    )
-
-
-def save(
-    data: TensorCollection,
-    prefix: str | None = None,
-    copy_existing: bool = False,
-    *,
-    num_threads: int = 0,
-    return_early: bool = False,
-    share_non_tensor: bool = False,
-    robust_key: bool | None = None,
-) -> None:
-    """Saves the tensordict to disk.
-
-    This function is a proxy to :meth:`~.memmap`.
-    """
-    return data.memmap(
-        prefix=prefix,
-        copy_existing=copy_existing,
-        num_threads=num_threads,
-        return_early=return_early,
-        share_non_tensor=share_non_tensor,
-        robust_key=robust_key,
-    )
-
-
-def memmap(
-    data: TensorCollection,
-    prefix: str | None = None,
-    copy_existing: bool = False,
-    *,
-    num_threads: int = 0,
-    return_early: bool = False,
-    share_non_tensor: bool = False,
-    robust_key: bool | None = None,
-) -> Self:
-    """Writes all tensors onto a corresponding memory-mapped Tensor in a new tensordict.
-
-    Args:
-        data (TensorDictBase): a data structure to save.
-        prefix (str): directory prefix where the memory-mapped tensors will
-            be stored. The directory tree structure will mimic the tensordict's.
-        copy_existing (bool): If False (default), an exception will be raised if an
-            entry in the tensordict is already a tensor stored on disk
-            with an associated file, but is not saved in the correct
-            location according to prefix.
-            If ``True``, any existing Tensor will be copied to the new location.
-
-    Keyword Args:
-        num_threads (int, optional): the number of threads used to write the memmap
-            tensors. Defaults to `0`.
-        return_early (bool, optional): if ``True`` and ``num_threads>0``,
-            the method will return a future of the tensordict.
-        share_non_tensor (bool, optional): if ``True``, the non-tensor data will be
-            shared between the processes and writing operation (such as inplace update
-            or set) on any of the workers within a single node will update the value
-            on all other workers. If the number of non-tensor leaves is high (e.g.,
-            sharing large stacks of non-tensor data) this may result in OOM or similar
-            errors. Defaults to ``False``.
-        robust_key (bool, optional): if ``True``, uses robust key encoding that safely
-            handles keys with path separators and special characters. If ``False``,
-            uses legacy behavior (keys used as-is). If ``None`` (default), emits a
-            deprecation warning and falls back to legacy behavior. Will default to
-            ``True`` in v0.12.
-
-    The TensorDict is then locked, meaning that any writing operations that
-    isn't in-place will throw an exception (eg, rename, set or remove an
-    entry).
-    Once the tensordict is unlocked, the memory-mapped attribute is turned to ``False``,
-    because cross-process identity is not guaranteed anymore.
-
-    Returns:
-        A new tensordict with the tensors stored on disk if ``return_early=False``,
-        otherwise a :class:`~tensordict.utils.TensorDictFuture` instance.
-
-    Note:
-        Serialising in this fashion might be slow with deeply nested tensordicts, so
-        it is not recommended to call this method inside a training loop.
-    """
-    return data.memmap(
-        prefix=prefix,
-        copy_existing=copy_existing,
-        num_threads=num_threads,
-        return_early=return_early,
-        share_non_tensor=share_non_tensor,
-        robust_key=robust_key,
-    )
+from tensordict._td_functions import (  # noqa: F401
+    cat,
+    from_consolidated,
+    from_module,
+    from_modules,
+    from_pytree,
+    fromkeys,
+    lazy_stack,
+    load,
+    load_memmap,
+    maybe_dense_stack,
+    memmap,
+    save,
+    stack,
+)
